合并了修改版本号8662-8664 从 branches/CodeTest: 合并了修改版本号8659-8661 从 branches/Code: 连锁掌柜-创意仓库 关注 日志记录 ........ 连锁掌柜-创意仓库 增加触点奖励 注册 关注 ........
........
合并了修改版本号8663 从 branches/Code:
连锁掌柜-创意仓库
........

........


git-svn-id: https://10.133.192.229/svn/CustomerManage/trunk/Code@8665 67192a19-bcbd-5648-baee-99b7ca3a6e28

+++ CPOS.DTO/Module/CreativityWarehouse/MarketingData/Response/GetLEventStatsRD.cs
@@ -15,11 +15,11 @@ namespace JIT.CPOS.DTO.Module.CreativityWarehouse.MarketingData.Response
         /// <summary>
         /// 现金券
         /// </summary>
-        public int CASHCouponValue { get; set; }
+        public decimal CASHCouponValue { get; set; }
         /// <summary>
         /// 折扣券
         /// </summary>
-        public int DISCOUNTCouponValue { get; set; }
+        public decimal DISCOUNTCouponValue { get; set; }
         /// <summary>
         /// 礼物券
         /// </summary>
